+++ sos21-database/migrations/20210520143144_affiliation_user_category.sql
@@ -1,0 +1,5 @@
+ALTER TABLE users
+    ALTER COLUMN affiliation DROP NOT NULL,
+    ADD CONSTRAINT users_affiliation_category CHECK (
+        (category = 'undergraduate_student') = (affiliation IS NOT NULL)
+    );