+++ sos21-database/src/model/user.rs
@@ -38,7 +38,7 @@ pub struct User {
     pub last_name: String,
     pub kana_last_name: String,
     pub phone_number: String,
-    pub affiliation: String,
+    pub affiliation: Option<String>,
     pub email: String,
     pub role: UserRole,
     pub category: UserCategory,