+++ sos21-domain/src/model/user/category.rs
@@ -1,6 +1,18 @@
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+use crate::model::user::UserAffiliation;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum UserCategory {
-    UndergraduateStudent,
+    UndergraduateStudent(UserAffiliation),
     GraduateStudent,
     AcademicStaff,
+}
+
+impl UserCategory {
+    pub fn affiliation(&self) -> Option<&UserAffiliation> {
+        match self {
+            UserCategory::UndergraduateStudent(affiliation) => Some(&affiliation),
+            UserCategory::GraduateStudent => None,
+            UserCategory::AcademicStaff => None,
+        }
+    }
 }