+++ sos21-gateway/database/src/user_repository.rs
@@ -82,7 +82,6 @@ fn from_user(user: User) -> data::user::User {
         kana_name,
         email,
         phone_number,
-        affiliation,
         role,
         category,
         assignment,
@@ -99,6 +98,15 @@ fn from_user(user: User) -> data::user::User {
         None => (None, None, None),
     };
 
+    let (category, affiliation) = match category {
+        UserCategory::UndergraduateStudent(affiliation) => (
+            data::user::UserCategory::UndergraduateStudent,
+            Some(affiliation.into_string()),
+        ),
+        UserCategory::GraduateStudent => (data::user::UserCategory::GraduateStudent, None),
+        UserCategory::AcademicStaff => (data::user::UserCategory::AcademicStaff, None),
+    };
+
     data::user::User {
         id: id.0,
         created_at: created_at.utc(),
@@ -108,18 +116,14 @@ fn from_user(user: User) -> data::user::User {
         kana_last_name,
         email: email.into_string(),
         phone_number: phone_number.into_string(),
-        affiliation: affiliation.into_string(),
+        affiliation,
         role: match role {
             UserRole::Administrator => data::user::UserRole::Administrator,
             UserRole::CommitteeOperator => data::user::UserRole::CommitteeOperator,
             UserRole::Committee => data::user::UserRole::Committee,
             UserRole::General => data::user::UserRole::General,
         },
-        category: match category {
-            UserCategory::UndergraduateStudent => data::user::UserCategory::UndergraduateStudent,
-            UserCategory::GraduateStudent => data::user::UserCategory::GraduateStudent,
-            UserCategory::AcademicStaff => data::user::UserCategory::AcademicStaff,
-        },
+        category,
         assignment: assignment.map(from_user_assignment),
         assignment_owner_project_id: owner_project_id.map(|id| id.to_uuid()),
         assignment_subowner_project_id: subowner_project_id.map(|id| id.to_uuid()),
@@ -179,6 +183,17 @@ pub fn to_user(user: data::user::User) -> Result<User> {
         None
     };
 
+    let category = match category {
+        data::user::UserCategory::UndergraduateStudent => {
+            let affiliation = affiliation
+                .context("category = 'undergraduate_student' but affiliation is null")?;
+            let affiliation = UserAffiliation::from_string(affiliation)?;
+            UserCategory::UndergraduateStudent(affiliation)
+        }
+        data::user::UserCategory::GraduateStudent => UserCategory::GraduateStudent,
+        data::user::UserCategory::AcademicStaff => UserCategory::AcademicStaff,
+    };
+
     Ok(User::from_content(UserContent {
         id: UserId(id),
         created_at: DateTime::from_utc(created_at),
@@ -186,18 +201,13 @@ pub fn to_user(user: data::user::User) -> Result<User> {
         kana_name: UserKanaName::from_string(kana_first_name, kana_last_name)?,
         email: UserEmailAddress::from_string(email)?,
         phone_number: PhoneNumber::from_string(phone_number)?,
-        affiliation: UserAffiliation::from_string(affiliation)?,
         role: match role {
             data::user::UserRole::Administrator => UserRole::Administrator,
             data::user::UserRole::CommitteeOperator => UserRole::CommitteeOperator,
             data::user::UserRole::Committee => UserRole::Committee,
             data::user::UserRole::General => UserRole::General,
         },
-        category: match category {
-            data::user::UserCategory::UndergraduateStudent => UserCategory::UndergraduateStudent,
-            data::user::UserCategory::GraduateStudent => UserCategory::GraduateStudent,
-            data::user::UserCategory::AcademicStaff => UserCategory::AcademicStaff,
-        },
+        category,
         assignment,
     }))
 }