+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Relax checkbox form item constraints ([#95](https://github.com/sohosai/sos21-backend/pull/95))
 - Enable to specify project by code when distributing files ([#96](https://github.com/sohosai/sos21-backend/pull/96))
+- Do not require affiliation of graduate students and academic staffs ([#93](https://github.com/sohosai/sos21-backend/pull/93))
 
 ### Deprecated
 ### Removed